Back:   se adiciona metodos  update y  delete

+++ PagosTemplate/Backend/PilaPayments/src/co/edu/uniandes/businesslogic/EntityLogic.java
@@ -45,6 +45,29 @@ public class EntityLogic {
 	 * Crear Entidad
 	 * @param entityTO
 	 */
+	public String update(PilaEntityTO entityTO){		
+		String response = "";		
+		PilaEntity entity = new PilaEntity();
+		entity.setCedula(entityTO.getCedula());
+		entity.setNombre(entityTO.getNombre());
+		entity.setApellido(entityTO.getApellido());
+		entity.setSalario(entityTO.getSalario());
+		entity.setTipoPension(entityTO.getTipoPension());
+		entity.setTipoPensionado(entityTO.getTipoPensionado());
+		entity.setActividad(entityTO.getActividad());
+		entity.setSuperEntidad(superEntityDAO.findSuperEntityById(entityTO.getSuperEntidad()));
+		entity.setPais(entityTO.getPais());
+		entity.setPaisGrupoFamiliar(entityTO.getPaisGrupoFamiliar());		
+		entity = entityDAO.update(entity);		
+		response = entity.getId().toString();		
+		return response;
+	}
+	
+	
+	/**
+	 * Crear Entidad
+	 * @param entityTO
+	 */
 	public String crearEntity(PilaEntityTO entityTO){
 		
 		String response = "";
@@ -68,28 +91,13 @@ public class EntityLogic {
 		return response;
 	}
 	
-	public String borrarEntity(PilaEntityTO entityTO){
-		
-		String response = "";
-		
-		PilaEntity entity = new PilaEntity();
-		entity.setCedula(entityTO.getCedula());
-		entity.setNombre(entity.getNombre());
-		entity.setApellido(entityTO.getApellido());
-		entity.setSalario(entityTO.getSalario());
-		entity.setTipoPension(entityTO.getTipoPension());
-		entity.setTipoPensionado(entityTO.getTipoPensionado());
-		entity.setActividad(entityTO.getActividad());
-		entity.setSuperEntidad(superEntityDAO.findSuperEntityById(entityTO.getSuperEntidad()));
-		entity.setPais(entityTO.getPais());
-		entity.setPaisGrupoFamiliar(entityTO.getPaisGrupoFamiliar());
-		
-		String id = entity.getId().toString();
-		
-		entityDAO.deleteEntity(entity);
-		
-		response = id;
-		
+	public String delete(Long id){		
+		String response = "";		
+		PilaEntity entity = entityDAO.findEntityById(id);
+		if(entity != null){
+			entityDAO.deleteEntity(entity);	
+			response = id+"";		
+		}
 		return response;
 	}
